refactor: replace ContentType switch with IBlockHandler registry (#10)
Split the ~75-line ContentType switch in PrinterService into one
IBlockHandler per block type, resolved from DI and dispatched by a
ContentType->handler dictionary. Adding a block type is now a new handler
class + registration, no switch to edit.

- Extract a pure BuildDocumentAsync (no I/O) from PrintAsync so the byte
  stream is buildable without the printer; PrintAsync keeps the status
  gate + network send.
- BlockContext threads the emitter, the mutable text encoding (CodePage
  blocks still switch it mid-document) and the HasCut flag.
- Shared CodePages (printer code page <-> .NET encoding) and StyledText
  (Text + Separator) helpers; per-handler mapping for barcode/QR.
- Handlers registered as stateless singletons to match PrinterService.

Verified live on the V330M via an A/B print (deployed master vs this
branch, identical document covering every block type, the mid-document
PC437 switch and the async image path). No unit tests by request.
